refactor: simplify QR code preview interaction tracking by centralizing event property construction and sanitization in a dedicated utility function
